+++ vsc-base.org/src/allAnnotations/annotations/TsIsValueAnnotatedCode.tsx
@@ -107,6 +107,7 @@ export const tsIsValue: (
    }
    return true
 }
+
 `}
       />
    )