+++ vsc-script/src/vsc-base-development/vsc-base-typescript.ts
@@ -556,7 +556,7 @@ export const tsCreateNodeVisitor = <T extends ts.Node = ts.SourceFile>(callback:
 `
 // Find a constant with name starting with 'module' within a function but not in an if statement
 const [_node, position] = vsc.tsFindNodePositionFromContent(source, node =>
- vsc.tsIsVariable(node, { 
+ vsc.tsMatchVariable(node, { 
       // test name of variable
       name: /^module/,
       // test if is in function
@@ -574,7 +574,7 @@ if (position) {
  * @returns [ts.Node | undefined, vsc.VscodePosition | undefined]
  */
 
-export const tsFindNodePositionFromContent = <TNode extends ts.Node = ts.Node>(source: string, callback: (node: ts.Node, typeChecker?: ts.TypeChecker, program?: ts.Program) => boolean, program?: ts.Program, fromPosition = 0, trimSpaces = true): [TNode | undefined, vsc.VscodePosition | undefined] => {
+export const tsFindNodePositionFromContent = <TNode extends ts.Node = ts.Node>(source: string, callback: (node: ts.Node, typeChecker?: ts.TypeChecker, program?: ts.Program) => TNode, program?: ts.Program, fromPosition = 0, trimSpaces = true): [TNode | undefined, vsc.VscodePosition | undefined] => {
    let position: vsc.VscodePosition | undefined
    let foundNode: TNode | undefined
    let typeChecker: ts.TypeChecker | undefined
@@ -586,15 +586,15 @@ export const tsFindNodePositionFromContent = <TNode extends ts.Node = ts.Node>(s
          if (node.pos < fromPosition) {
             return ts.visitEachChild(node, (child) => visit(child), context);
          }
-         const found = callback(node, typeChecker, program);
-         if (!found) {
+         const _foundNode = callback(node, typeChecker, program);
+         if (!_foundNode) {
             return ts.visitEachChild(node, (child) => visit(child), context);
          }
          if (node === undefined) {
             throw new Error('Node is undefined!!!')
          }
          position = vsc.createVscodeRangeAndPosition(source, node.pos, node.end, trimSpaces);
-         foundNode = node as TNode
+         foundNode = _foundNode
          return node
       }
       return (node) => ts.visitNode(node, visit);
@@ -624,7 +624,7 @@ export const tsFindNodePositionFromContent = <TNode extends ts.Node = ts.Node>(s
 `
 // Find a constant with name starting with 'module' within a function but not in an if statement
 const nodePositionArray = vsc.tsFindAllNodePositionsFromContent(source, node =>
- vsc.tsIsVariable(node, { 
+ vsc.tsMatchVariable(node, { 
       // test name of variable
       name: /^module/,
       // test if is in function
@@ -640,10 +640,8 @@ nodePositionArray.forEach([node, position] => {
  * @returns [ts.Node, vsc.VscodePosition][]
  */
 
-export const tsFindAllNodePositionsFromContent = <TNode extends ts.Node = ts.Node>(source: string, callback: (node: ts.Node, typeChecker?: ts.TypeChecker, program?: ts.Program) => boolean, program?: ts.Program, fromPosition = 0, trimSpaces = true): [TNode, vsc.VscodePosition][] => {
+export const tsFindAllNodePositionsFromContent = <TNode extends ts.Node = ts.Node>(source: string, callback: (node: ts.Node, typeChecker?: ts.TypeChecker, program?: ts.Program) => TNode, program?: ts.Program, fromPosition = 0, trimSpaces = true): [TNode, vsc.VscodePosition][] => {
    let positions: [TNode, vsc.VscodePosition][] = [];
-   let position: vsc.VscodePosition
-   let foundNode: TNode
    let typeChecker: ts.TypeChecker | undefined
    if (program) {
       typeChecker = program.getTypeChecker()
@@ -653,16 +651,15 @@ export const tsFindAllNodePositionsFromContent = <TNode extends ts.Node = ts.Nod
          if (node.pos < fromPosition) {
             return ts.visitEachChild(node, (child) => visit(child), context);
          }
-         const found = callback(node, typeChecker, program);
-         if (!found) {
+         const _foundNode = callback(node, typeChecker, program);
+         if (!_foundNode) {
             return ts.visitEachChild(node, (child) => visit(child), context);
          }
          if (node === undefined) {
             throw new Error('Node is undefined!!!')
          }
-         position = vsc.createVscodeRangeAndPosition(source, node.pos, node.end, trimSpaces);
-         foundNode = node as TNode;
-         positions.push([foundNode, position])
+         const position = vsc.createVscodeRangeAndPosition(source, node.pos, node.end, trimSpaces);
+         positions.push([_foundNode, position])
          return ts.visitEachChild(node, (child) => visit(child), context);
       }
       return (node) => ts.visitNode(node, visit);
@@ -688,7 +685,7 @@ let source = `
    }
 `
 // Find a constant with name starting with 'module' within a function but not in an if statement
-source = vsc.tsReplace(source, '/module/area/file2', node => vsc.tsIsValue(node, /\/area\/file1/, {
+source = vsc.tsReplace(source, '/module/area/file2', node => vsc.tsMatchValue(node, /\/area\/file1/, {
    hasAncestors: [
       ancestor => vsc.tsIsFunction(ancestor, { name: /^method/ }),
       ancestor => vsc.tsIsVariable(ancestor, { name: /^module.*Path/ })
@@ -697,7 +694,7 @@ source = vsc.tsReplace(source, '/module/area/file2', node => vsc.tsIsValue(node,
  *
  * @returns string
  */
-export const tsReplace = (source: string, replaceString: string, callback: (node: ts.Node, typeChecker?: ts.TypeChecker, program?: ts.Program) => boolean, program?: ts.Program, fromPosition = 0, trimSpaces = true): string => {
+export const tsReplace = (source: string, replaceString: string, callback: (node: ts.Node, typeChecker?: ts.TypeChecker, program?: ts.Program) => ts.Node, program?: ts.Program, fromPosition = 0, trimSpaces = true): string => {
    const [node, position] = vsc.tsFindNodePositionFromContent(source, callback, program, fromPosition, trimSpaces);
    if (position) {
       //replace
@@ -720,13 +717,13 @@ let source = `
    }
 `
 // Find a constant with name starting with 'module' within a function but not in an if statement
-source = vsc.tsReplaceAll(source, 'moduleNumber2', node => vsc.tsIsIdentifier(node, {
+source = vsc.tsReplaceAll(source, 'moduleNumber2', node => vsc.tsMatchIdentifier(node, {
    name: 'moduleNumber1Path'
 }))
  *
  * @returns string
  */
-export const tsReplaceAll = (source: string, replaceString: string, callback: (node: ts.Node, typeChecker?: ts.TypeChecker, program?: ts.Program) => boolean, program?: ts.Program, fromPosition = 0, trimSpaces = true): string => {
+export const tsReplaceAll = (source: string, replaceString: string, callback: (node: ts.Node, typeChecker?: ts.TypeChecker, program?: ts.Program) => ts.Node, program?: ts.Program, fromPosition = 0, trimSpaces = true): string => {
    const positions = vsc.tsFindAllNodePositionsFromContent(source, callback, program, fromPosition, trimSpaces);
    positions.sort(([, positionA], [, positionB]) => positionA.start - positionB.start)
    let diff = 0
@@ -1566,7 +1563,7 @@ export const tsIsEnumMember: (node: ts.Node | undefined, options?: {
  * @oneLineEx const found = vsc.tsIsNode(node, options)
  * @ex 
 const found = vsc.tsIsNode(node, { name: /^keyName$/ })
- * @returns ts.PropertyAssignment | undefined
+ * @returns boolean
  */
 export const tsIsNode: (node: ts.Node | undefined, options?: {
    name?: RegExp | string
@@ -1637,7 +1634,6 @@ const foundNumberExpression = vsc.tsIsValue(node, 12)
 const foundNumberExpression = vsc.tsIsValue(node, 12, {
    hasParent: parent => vsc.matchEnum(parent)
 })
-
  * @returns boolean
  */
 export const tsIsValue: (
@@ -1704,4 +1700,70 @@ export const tsIsValue: (
       return false
    }
    return true
+}
+
+
+/** vsc-base method
+ * @description
+ * Base test for node properties. \
+ * Optional test for its name with a string or regexp. \
+ * (return false for node that don't have name property)\
+ * Optional test for tsHasAncestor and hasGrandChild \
+ * See [tsHasAncestor](http://vsc-base.org/#tsHasAncestor), [tsHasAncestors](http://vsc-base.org/#tsHasAncestors), [hasGrandChild](http://vsc-base.org/#hasGrandChild) and [hasGrandChildren](http://vsc-base.org/#hasGrandChildren) \
+ * Optional value can be tested against a string, a number (with a string, number or regexp). \
+ * (return false for node that don't have initializer)\
+ * See [tsIsValue](http://vsc-base.org/#tsIsValue) \
+ * @see [tsMatchNode](http://vsc-base.org/#tsMatchNode)
+ * @vscType ts
+ * @oneLineEx const foundNode = vsc.tsMatchNode(node, options)
+ * @ex 
+const foundNode = vsc.tsMatchNode(node, { name: /^keyName$/ })
+ * @returns s.Node | undefined
+ */
+export const tsMatchNode: (node: ts.Node | undefined, options?: {
+   name?: RegExp | string
+   value?: (RegExp | string | number | boolean | null)
+   hasParent?: (parent: ts.Node) => boolean
+   hasAncestor?: (parent: ts.Node, depth: number) => boolean
+   hasAncestors?: ((parent: ts.Node, depth: number) => boolean)[]
+   hasGrandChild?: (child: ts.Node, depth: number) => boolean
+   hasGrandChildren?: ((child: ts.Node, depth: number) => boolean)[]
+}) => ts.Node | undefined = (node, options) => {
+   if (vsc.tsIsNode(node, options)) {
+      return node
+   }
+   return undefined
+}
+
+/** vsc-base method
+ * @description
+ * Test if node is an value: string expression, number expression or boolean (true or false) \
+ * and match the value: true, false, a number, a string, \
+ * A RegExp can be applied for string/number search. \
+ * Optional test hasAncestor. \
+ * See [tsHasAncestor](http://vsc-base.org/#tsHasAncestor) and [tsHasAncestors](http://vsc-base.org/#tsHasAncestors)
+ * @see [tsMatchValueNode](http://vsc-base.org/#tsMatchValueNode)
+ * @vscType ts
+ * @oneLineEx const foundNode = vsc.tsMatchValueNode(node, value)
+ * @ex 
+// Found a NumberExpression with value 12
+const foundNode = vsc.tsMatchValueNode(node, 12)
+// Found a NumberExpression with value 12, with a parent EnumValue
+const foundNode = vsc.tsMatchValueNode(node, 12, {
+   hasParent: parent => vsc.matchEnum(parent)
+})
+ * @returns s.Node | undefined
+ */
+export const tsMatchValueNode: (
+   node: ts.Node | undefined,
+   value: (RegExp | string | number | boolean | null),
+   options?: {
+      hasParent: (parent: ts.Node) => boolean
+      hasAncestor?: (parent: ts.Node, depth: number) => boolean
+      hasAncestors?: ((parent: ts.Node, depth: number) => boolean)[]
+   }
+) => ts.Node | undefined = (node, matchValue, options) => {
+   if (vsc.tsIsValue(node, matchValue, options)) {
+      return node
+   }
 }